Refactor: 답장 컴포넌트 분리
