+++ src/components/Messages/Message.jsx
@@ -36,6 +36,7 @@ const UserInfo = styled.div`
     > div {
         margin-bottom: 10px;
     }
+    white-space: pre-wrap;
 `;
 
 const ProfileImage = styled.img`
