+++ src/components/Messages/index.jsx
@@ -6,17 +6,11 @@ import MessageRemove from './MessageRemove';
 import { replyInput } from 'store';
 import styled from 'styled-components';
 import { useRef } from 'react';
+import MessageReply from './MessageReply';
 
 const Messages = (props) => {
     const dispatch = useDispatch();
     const data = useSelector((state) => state.message);
-    const handleReply = (event) => {
-        const selectMessage = data.filter(
-            (item) => item.commentId === event.target.id
-        )[0];
-        const input = `${selectMessage.userName}\n${selectMessage.content}\n(회신)\n`;
-        dispatch(replyInput(input));
-    };
 
     return (
         <StyledMessages ref={props.MsgBox}>
@@ -33,13 +27,7 @@ const Messages = (props) => {
                             {item.userName}
                         </Message>
                         <ButtonBox className="button_box">
-                            <ReplyButton
-                                className="reply_div"
-                                onClick={handleReply}
-                                id={item.commentId}
-                                type="button"
-                                value="답장"
-                            />
+                            <MessageReply id={item.commentId} />
                             <MessageRemove
                                 id={item.commentId}
                                 content={item.content}
@@ -84,10 +72,6 @@ const ButtonBox = styled.div`
         border: none;
         box-shadow: 2px 2px 2px 1px rgba(0, 0, 255, 0.2);
     }
-`;
-
-const ReplyButton = styled.input`
-    margin-right: 5px;
 `;
 
 export default Messages;
